Properly version the CSS when served under DEBUG mode
CSS is being served without content hash in the URL so that the
LiveReload plugin can detect changes to the file. However, this
causes the browser to cache the CSS files incorrectly, which
is pretty annoying for development.

Added a hash of the CSS file contents so that it gets refreshed on loads